+++ modules/apim-adapter/src/main/java/com/axway/apim/api/apiSpecification/WADLSpecification.java
@@ -24,21 +24,21 @@ public class WADLSpecification extends APISpecification {
 	}
 
 	@Override
-	public void configureBasepath(String backendBasepath, API api) throws AppException {
+	public void configureBasePath(String backendBasePath, API api) throws AppException {
 		if(!CoreParameters.getInstance().isReplaceHostInSwagger()) return;
 		try {
-			if(backendBasepath!=null) {
-				URL url = new URL(backendBasepath); // Parse it to make sure it is valid
-				if(url.getPath()!=null && !url.getPath().equals("") && !backendBasepath.endsWith("/")) { // See issue #178
-					backendBasepath += "/";
+			if(backendBasePath!=null) {
+				URL url = new URL(backendBasePath); // Parse it to make sure it is valid
+				if(url.getPath()!=null && !url.getPath().equals("") && !backendBasePath.endsWith("/")) { // See issue #178
+					backendBasePath += "/";
 				}
 				// The WADL has the base path configured like so: <resources base="http://customer-api.ddns.net:8099/">
-				wadl = wadl.replaceFirst("(<resources.*base=\").*(\">)", "$1"+backendBasepath+"$2");
+				wadl = wadl.replaceFirst("(<resources.*base=\").*(\">)", "$1"+backendBasePath+"$2");
 
 				this.apiSpecificationContent = wadl.getBytes();
 			}
 		} catch (MalformedURLException e) {
-			throw new AppException("The configured backendBasepath: '"+backendBasepath+"' is invalid.", ErrorCode.BACKEND_BASEPATH_IS_INVALID, e);
+			throw new AppException("The configured backendBasePath: '"+backendBasePath+"' is invalid.", ErrorCode.BACKEND_BASEPATH_IS_INVALID, e);
 		} catch (Exception e) {
 			LOG.error("Cannot replace host in provided Swagger-File. Continue with given host.", e);
 		}
@@ -55,7 +55,7 @@ public class WADLSpecification extends APISpecification {
 			LOG.debug("No WADL specification. Specification doesn't contain WADL namespace: wadl.dev.java.net in the first 500 characters.");
 			return false;
 		}
-			// We going to use a cheap way - Avoid parsing & writing back the WADL-File.
+			// We are going to use a cheap way - Avoid parsing & writing back the WADL-File.
 		this.wadl = new String(apiSpecificationContent);
 		return true;
 	}