+++ modules/apis/src/test/java/com/axway/apim/api/definition/APISpecificationSwagger1xTest.java
@@ -32,7 +32,7 @@ public class APISpecificationSwagger1xTest {
 	public void standardPort() throws AppException, IOException {
 		byte[] content = getSwaggerContent(testPackage + "/swagger12.json");
 		APISpecification apiDefinition = APISpecificationFactory.getAPISpecification(content, "teststore.json", "Test-API");
-		apiDefinition.configureBasepath("https://petstore.swagger.io", null);
+		apiDefinition.configureBasePath("https://petstore.swagger.io", null);
 		
 		Assert.assertTrue(apiDefinition instanceof Swagger1xSpecification, "Specification must be an Swagger12Specification");
 		Assert.assertEquals(apiDefinition.getDescription(), "Swagger 1.2 Description");
@@ -44,7 +44,7 @@ public class APISpecificationSwagger1xTest {
 	public void specificPort() throws AppException, IOException {
 		byte[] content = getSwaggerContent(testPackage + "/swagger12.json");
 		APISpecification apiDefinition = APISpecificationFactory.getAPISpecification(content, "teststore.json", "Test-API");
-		apiDefinition.configureBasepath("https://petstore.swagger.io:8180", null);
+		apiDefinition.configureBasePath("https://petstore.swagger.io:8180", null);
 		
 		Assert.assertTrue(apiDefinition instanceof Swagger1xSpecification, "Specification must be an Swagger12Specification");
 		JsonNode swagger = mapper.readTree(apiDefinition.getApiSpecificationContent());
@@ -55,7 +55,7 @@ public class APISpecificationSwagger1xTest {
 	public void standardPortGiven() throws AppException, IOException {
 		byte[] content = getSwaggerContent(testPackage + "/swagger12.json");
 		APISpecification apiDefinition = APISpecificationFactory.getAPISpecification(content, "teststore.json", "Test-API");
-		apiDefinition.configureBasepath("https://petstore.swagger.io:443", null);
+		apiDefinition.configureBasePath("https://petstore.swagger.io:443", null);
 		
 		Assert.assertTrue(apiDefinition instanceof Swagger1xSpecification, "Specification must be an Swagger12Specification");
 		JsonNode swagger = mapper.readTree(apiDefinition.getApiSpecificationContent());
@@ -66,7 +66,7 @@ public class APISpecificationSwagger1xTest {
 	public void includingBasePath() throws AppException, IOException {
 		byte[] content = getSwaggerContent(testPackage + "/swagger12.json");
 		APISpecification apiDefinition = APISpecificationFactory.getAPISpecification(content, "teststore.json", "Test-API");
-		apiDefinition.configureBasepath("https://petstore.swagger.io:443/myapi", null);
+		apiDefinition.configureBasePath("https://petstore.swagger.io:443/myapi", null);
 		
 		Assert.assertTrue(apiDefinition instanceof Swagger1xSpecification, "Specification must be an Swagger12Specification");
 		JsonNode swagger = mapper.readTree(apiDefinition.getApiSpecificationContent());
@@ -77,19 +77,19 @@ public class APISpecificationSwagger1xTest {
 	public void testSwagger11Specification() throws AppException, IOException {
 		byte[] content = getSwaggerContent(testPackage + "/swagger11.json");
 		APISpecification apiDefinition = APISpecificationFactory.getAPISpecification(content, "teststore.json", "Test-API");
-		apiDefinition.configureBasepath("https://petstore.swagger.io:443/myapi/", null);
+		apiDefinition.configureBasePath("https://petstore.swagger.io:443/myapi/", null);
 		
 		Assert.assertTrue(apiDefinition instanceof Swagger1xSpecification, "Specification must be an Swagger12Specification");
 		JsonNode swagger = mapper.readTree(apiDefinition.getApiSpecificationContent());
 		Assert.assertEquals(swagger.get("basePath").asText(), "https://petstore.swagger.io:443/myapi/");
 	}
 	
-	@Test(expectedExceptions = AppException.class, expectedExceptionsMessageRegExp = "The configured backendBasepath: 'An-Invalid-URL' is invalid.")
+	@Test(expectedExceptions = AppException.class, expectedExceptionsMessageRegExp = "The configured backendBasePath: 'An-Invalid-URL' is invalid.")
 	public void testInvalidBackendBasepath() throws AppException, IOException {
 
 		byte[] content = getSwaggerContent(testPackage + "/swagger11.json");
 		APISpecification apiDefinition = APISpecificationFactory.getAPISpecification(content, "teststore.json", "Test-API");
-		apiDefinition.configureBasepath("An-Invalid-URL", null);
+		apiDefinition.configureBasePath("An-Invalid-URL", null);
 	}
 	
 	private byte[] getSwaggerContent(String swaggerFile) throws AppException {